Move code out of name step and make Git prompts first.  SWA named after owner-repo
